Move toast related files to src/scripts/utils/toast
